cyber/requirements.txt
# Core dependencies
networkx==3.2.1
neo4j==5.14.1
pandas==2.1.4
numpy==1.26.2
requests==2.31.0
python-dateutil==2.8.2

# Parsing
xmltodict==0.13.0
lxml==4.9.4
beautifulsoup4==4.12.2

# Reporting
reportlab==4.0.7
jinja2==3.1.2
matplotlib==3.8.2
plotly==5.18.0

# ML/Stats
scikit-learn==1.3.2
scipy==1.11.4

# Utilities
pyyaml==6.0.1
python-dotenv==1.0.0
colorama==0.4.6
tqdm==4.66.1

# Logging
loguru==0.7.2
